+++ src/main/java/nextstep/exception/ExceptionMessage.java
@@ -1,11 +1,9 @@
-package nextstep.subway.exception;
+package nextstep.exception;
 
 import lombok.Getter;
 
 @Getter
 public enum ExceptionMessage {
-//    UPSTATION_VALIDATION_EXCEPTION("새로운 구간의 상행역은 해당 노선에 등록되어있는 하행 종점역이어야 합니다."),
-//    DOWNSTATION_VALIDATION_EXCEPTION("이미 해당 노선에 등록되어있는 역은 새로운 구간의 하행역이 될 수 없습니다."),
     NEW_SECTION_VALIDATION_EXCEPTION("새로운 구간의 상행역과 하행역은 같을 수 없습니다."),
     DELETE_LAST_SECTION_EXCEPTION("마지막 종착역만 제거할 수 있습니다."),
     DELETE_ONLY_ONE_SECTION_EXCEPTION("지하철 노선에 구간이 1개인 경우 역을 삭제할 수 없다."),
@@ -16,7 +14,8 @@ public enum ExceptionMessage {
     LONGER_DISTANCE_SECTION_EXCEPTION("기존 구간보다 거리가 긴 구간은 추가할 수 없습니다."),
     SAME_SOURCE_TARGET_EXCEPTION("출발역과 도착역이 같습니다."),
     NOT_CONNECTED_EXCEPTION("출발역과 도착역이 연결되어 있지 않습니다."),
-    NO_EXISTS_STATION_EXCEPTION("출발역 또는 도착역이 존재하지 않습니다.");
+    NO_EXISTS_STATION_EXCEPTION("존재하지 않는 지하철역 입니다."),
+    NO_EXISTS_MEMBER_EXCEPTION("존재하지 않는 회원 입니다.");
 
     private final String message;
