+++ src/main/java/nextstep/exception/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.exception;
 
 import lombok.Builder;
 import lombok.Getter;
