+++ src/main/java/nextstep/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
