+++ src/main/java/nextstep/subway/service/SectionService.java
@@ -5,8 +5,8 @@ import nextstep.subway.domain.entity.Section;
 import nextstep.subway.domain.entity.Sections;
 import nextstep.subway.domain.entity.Station;
 import nextstep.subway.domain.response.SectionResponse;
-import nextstep.subway.exception.ApplicationException;
-import nextstep.subway.exception.ExceptionMessage;
+import nextstep.exception.ApplicationException;
+import nextstep.exception.ExceptionMessage;
 import nextstep.subway.repository.LineRepository;
 import nextstep.subway.repository.SectionRepository;
 import org.springframework.stereotype.Service;
