+++ src/test/java/nextstep/subway/acceptance/PathAcceptanceTest.java
@@ -3,7 +3,7 @@ package nextstep.subway.acceptance;
 import nextstep.subway.domain.request.LineRequest;
 import nextstep.subway.domain.response.PathResponse;
 import nextstep.subway.domain.response.StationResponse;
-import nextstep.subway.exception.ExceptionResponse;
+import nextstep.exception.ExceptionResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,7 +13,7 @@ import org.springframework.test.context.jdbc.Sql;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static nextstep.subway.exception.ExceptionMessage.*;
+import static nextstep.exception.ExceptionMessage.*;
 import static nextstep.subway.utils.LineTestUtil.지하철_노선_생성;
 import static nextstep.subway.utils.PathTestUtil.getShortestPath;
 import static nextstep.subway.utils.SectionTestUtil.지하철_구간_추가;
