+++ src/test/java/nextstep/subway/unit/PathServiceMockTest.java
@@ -4,8 +4,8 @@ import nextstep.subway.domain.entity.Section;
 import nextstep.subway.domain.entity.Station;
 import nextstep.subway.domain.response.PathResponse;
 import nextstep.subway.domain.response.StationResponse;
-import nextstep.subway.exception.ApplicationException;
-import nextstep.subway.exception.ExceptionMessage;
+import nextstep.exception.ApplicationException;
+import nextstep.exception.ExceptionMessage;
 import nextstep.subway.repository.SectionRepository;
 import nextstep.subway.service.PathService;
 import nextstep.subway.service.StationService;
